requirements.txt
# CORE, PROVEN, STABLE VERSIONS FOR APRIL24TH SUPERLIQUID
finrl==0.3.4                # Last stable for SB3 1.7.0 and gym 0.21.0
stable-baselines3==1.7.0    # Compatible with finrl 0.3.4
# gym==0.21.0               # Installed from pre-built wheel in Dockerfile
gymnasium==0.26.3           # Used by FinRL, but not by SB3
torch==2.0.1+cu118          # CUDA 11.8 wheel
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.3.0
ccxt==3.1.47
matplotlib==3.7.2
# Add only proven, compatible extras below
fastapi==0.110.2
uvicorn[standard]==0.34.2
psycopg2-binary==2.9.9
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.11.4
PyJWT==2.10.1
ray==2.45.0
celery==5.3.6
httpx==0.27.0
prometheus_client==0.21.1
pyarrow==20.0.0
sqlalchemy==2.0.30
gradio==4.26.0
rq==1.16.1
aioredis==2.0.1
pytest==8.3.5
hypothesis==6.100.0
pyzmq==26.0.3
databento
transformers==4.41.2
boto3==1.34.113
PyGithub==2.3.0
prometheus-fastapi-instrumentator==6.1.0
slowapi==0.1.9
